+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.3.1
 
 - Add script abstraction
-- Add zmscore command
+- Add zmscore, getdel command
 
 ## 1.3.0